refactor: separate RealWorld APIs and Supabase APIs from Next.js API directory structure
