+++ src/components/article/CommentsList.tsx
@@ -1,6 +1,6 @@
 import Link from 'next/link';
 import Image from 'next/image';
-import { syncCommentsWithSupabase, fetchCommentsFromSupabase } from '@/app/api/supabase';
+import { syncCommentsWithSupabase, fetchCommentsFromSupabase } from '@/dataSources/supabase';
 import { Comments } from '@/types';
 import { CardFooterProps, ProfileLinkProps } from './type';
 import { formatDate } from '@/utils/format';
