+++ src/dataSources/supabase/sync/syncArticles.ts
@@ -1,5 +1,5 @@
 import { supabase } from '@/services/supabaseClient';
-import { fetchArticlesFromRealworld } from '@/app/api/realworld';
+import { fetchArticlesFromRealworld } from '@/dataSources/realworld';
 import { ArticlesApiParam } from '@/types';
 
 // Supabase와 동기화하는 함수 : RealWorld API에서 데이터를 가져와 Supabase에 저장하고, 이미 존재하는 데이터는 중복 삽입을 방지
